@@ -1,9 +1,10 @@
 package com.MaFederation.MaFederation.dto.Category;
 
 public record CategoryDTO(
-    Integer categoryId,
+    Integer id,
     String name,
     String description,
     Integer ageMin,
     Integer ageMax
+    
 ) {}
