@@ -3,13 +3,15 @@ package com.MaFederation.MaFederation.dto.Club;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ResponseClubDTO {
-    Integer clubId;
+    Integer id;
     private String name;
     private String location;
     private Integer foundedYear;
@@ -21,4 +23,9 @@ public class ResponseClubDTO {
     private List<Integer> categoryIds;
     private List<Integer> memberIds;
     private List<Integer> fileIds;
+    private LocalDateTime createdAt;
+    private LocalDateTime updatedAt;
+    private String createdBy;
+    private String updatedBy;
+
 }
