@@ -1,46 +1,66 @@
 package com.MaFederation.MaFederation.mappers;
 
+import com.MaFederation.MaFederation.model.User;
+import com.MaFederation.MaFederation.model.UserFile;
 import com.MaFederation.MaFederation.dto.User.ResponseUserDTO;
 import com.MaFederation.MaFederation.dto.User.UserPostDTO;
-import com.MaFederation.MaFederation.model.User;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
-@Service
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Component
 public class UserMapper {
 
     public ResponseUserDTO toResponseDto(User user) {
         if (user == null) return null;
 
-        return new ResponseUserDTO(
-            user.getUserId(),
-            user.getProfilePicture(),
-            user.getEmail(),
-            user.getFirstName(),
-            user.getLastName(),
-            user.getDateOfBirth(),
-            user.getGender(),
-            user.getPhoneNumber(),
-            user.getAddress(),
-            user.getNationalID(),
-            user.getNationality()
-        );
+        ResponseUserDTO dto = new ResponseUserDTO();
+        dto.setId(user.getId());
+        dto.setProfilePicture(user.getProfilePicture());
+        dto.setEmail(user.getEmail());
+        dto.setFirstName(user.getFirstName());
+        dto.setLastName(user.getLastName());
+        dto.setDateOfBirth(user.getDateOfBirth());
+        dto.setGender(user.getGender());
+        dto.setPhoneNumber(user.getPhoneNumber());
+        dto.setAddress(user.getAddress());
+        dto.setNationalID(user.getNationalID());
+        dto.setNationality(user.getNationality());
+        dto.setType(user.getType());
+        dto.setCreatedAt(user.getCreatedAt());
+        dto.setUpdatedAt(user.getUpdatedAt());
+        dto.setCreatedBy(user.getCreatedBy());
+        dto.setUpdatedBy(user.getUpdatedBy());
+
+        List<Integer> fileIds = user.getFiles() != null
+            ? user.getFiles().stream()
+                .map(UserFile::getId)
+                .collect(Collectors.toList())
+            : null;
+
+        dto.setFileIds(fileIds);
+
+        return dto;
     }
 
     public User toEntity(UserPostDTO dto) {
         if (dto == null) return null;
 
-        return User.builder()
-                .profilePicture(dto.getProfilePicture())
-                .passwordHash(dto.getPasswordHash())
-                .email(dto.getEmail())
-                .firstName(dto.getFirstName())
-                .lastName(dto.getLastName())
-                .dateOfBirth(dto.getDateOfBirth())
-                .gender(dto.getGender())
-                .phoneNumber(dto.getPhoneNumber())
-                .address(dto.getAddress())
-                .nationalID(dto.getNationalID())
-                .nationality(dto.getNationality())
-                .build();
+        User user = new User();
+        user.setProfilePicture(dto.getProfilePicture());
+        user.setPasswordHash(dto.getPasswordHash());
+        user.setEmail(dto.getEmail());
+        user.setFirstName(dto.getFirstName());
+        user.setLastName(dto.getLastName());
+        user.setDateOfBirth(dto.getDateOfBirth());
+        user.setGender(dto.getGender());
+        user.setPhoneNumber(dto.getPhoneNumber());
+        user.setAddress(dto.getAddress());
+        user.setNationalID(dto.getNationalID());
+        user.setNationality(dto.getNationality());
+        user.setType(dto.getType());
+
+        return user;
     }
 }
