@@ -1,26 +1,26 @@
 package com.MaFederation.MaFederation.model;
 
-import com.MaFederation.MaFederation.enums.PlayerFileType;
+import com.MaFederation.MaFederation.enums.UserFileType;
 import jakarta.persistence.*;
 import lombok.*;
+import lombok.experimental.SuperBuilder;
 
 @Entity
 @Table(name = "user_files")
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
-public class UserFile {
+@SuperBuilder
+@EqualsAndHashCode(callSuper = true)
+public class UserFile extends Audit {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    
     @Lob
     private byte[] content;
     @Enumerated(EnumType.STRING)
 
     
-    private PlayerFileType type;
+    private UserFileType type;
 
     @ManyToOne
     @JoinColumn(name = "user_id", nullable = false)
