@@ -2,8 +2,7 @@
 import { UserResponse } from "../User/userResponse";
 
 export interface ClubMemberResponse extends UserResponse {
-    clubId: number; 
-    type: string;
+    clubId: number;
     categories: string[];
 }
 
