@@ -1,4 +1,4 @@
-export type FileType = 'PROFILE_PICTURE' | 'PASSPORT' | 'CV' | 'LICENSE' | 'OTHER';
+export type FileType = 'PASSPORT' | 'CV' | 'LICENSE' | 'OTHER';
 
 export interface UserPostFile {
   fileUrl: string;
